tidy_lm_bootstrap_permute

permute lm and output the results as a tidy table.

tidy_lm_bootstrap_permute(data, formula, n_bootstrap = 1000,
  bootstrap_ci = 0.95, n_permute = 1000, var_permute,
  r_squared_ci = 0.95)

Arguments

data
a data.frame to be analyzed

formula
a formula to be analyzed as typically written for the lm function

n_bootstrap
= 1000 (default) the number of bootstraps to perform

bootstrap_ci
= 0.95 (default)

n_permute
= 1000 (default) the number of permutations to perform

var_permute
variable(s) to unlink in the permutation

r_squared_ci
= 0.95 (default)

Value

outputs tidy table that includes the p.value from the permutation of a lm test

Examples

packages <- c("broom", "broomExtra", "dplyr", "modelr", "purrr", "tibble")
xfun::pkg_attach2(packages, message = F)

data <- tibble(
  a = scale(sample.int(100), scale = F),
  b = scale(sample.int(100), scale = F),
  c = b^2,
  d = scale(sample.int(100), scale = F)
)

tidy_lm_bootstrap_permute(data, a ~ b, var_permute = "a")

#> Source: local data frame [2 x 14]
#> Groups: <by row>
#> 
#> # A tibble: 2 x 14
#>   term   estimate std.error statistic p.value estimate.ci.0.9… estimate.ci.0.9…
#>   <chr>     <dbl>     <dbl>     <dbl>   <dbl>            <dbl>            <dbl>
#> 1 (Int… -1.74e-16     2.92  -5.95e-17   1               -5.85             6.09 
#> 2 b     -1.92e- 3     0.101 -1.90e- 2   0.985           -0.203            0.219
#> # … with 7 more variables: p.permute.1000 <dbl>,
#> #   r.squared.ci.0.95.ll.boot.1000 <dbl>, r.squared.ci.0.95.ul.boot.1000 <dbl>,
#> #   adj.r.squared.ci.0.95.ll.boot.1000 <dbl>,
#> #   adj.r.squared.ci.0.95.ul.boot.1000 <dbl>, r.squared <dbl>,
#> #   adj.r.squared <dbl>